# docs/swagger.yaml
basePath: /v1
...
host: localhost:5000 devel
info:
  contact:
    name: API Support
  description: Description for what is this security definition being used
  license:
    name: GNU GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
  termsOfService: http://swagger.io/terms/
  title: GO API template
  version: "1.0"
paths:
  /auth/login:
    get:
      consumes:
      - application/json
      description: Get your user's token to acess users only protected routes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
      summary: login to get the authentication bearer token
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"